requirements.txt
cloudinary==1.21.1
flask==1.1.2
Flask_SQLAlchemy==2.4.3
gunicorn==20.0.4
passlib==1.7.2
psycopg2-binary==2.8.5
pyjwt[crypto]==1.7.1
requests==2.24.0
